p2p/discover: Relocate test helpers
