+++ networks/p2p/discover/table_init_test.go
@@ -72,54 +72,3 @@ func Test_Table_StartClose(t *testing.T) {
 	tab.Close() // idempotent
 }
 
-//// Test helpers
-
-// newTestNode generates a cryptographically valid Node with a loopback IP.
-// The loopback IP is treated as LAN and bypasses the kademliaStorage IP limit.
-func newTestNode(t *testing.T, nType NodeType) *Node {
-	t.Helper()
-	key, err := crypto.GenerateKey()
-	require.NoError(t, err)
-	return NewNode(PubkeyID(&key.PublicKey), net.IP{127, 0, 0, 1}, 30303, 30303, nil, nType)
-}
-
-// newTestTable2 constructs a Table2 backed by an in-memory node database and
-// a no-op UDP transport. tab.Close is registered as a test cleanup, so callers
-// do not have to call it explicitly (though doing so a second time is safe).
-func newTestTable2(t *testing.T, transport transport) *Table2 {
-	t.Helper()
-	cfg := &Config{
-		Id:         NodeID{},
-		Addr:       &net.UDPAddr{IP: net.IP{127, 0, 0, 1}, Port: 30303},
-		NodeType:   NodeTypeEN,
-		NodeDBPath: "", // In-memory temporary database
-	}
-	if transport == nil {
-		transport = &noopTransport{}
-	}
-	tab, err := newTable2(cfg, transport)
-	require.NoError(t, err)
-	t.Cleanup(tab.Close)
-	return tab
-}
-
-// newTestBootnode generates a cryptographically valid Node at the given IP,
-// suitable for use as a bootstrap node.
-func newTestBootnode(t *testing.T, ip net.IP) *Node {
-	t.Helper()
-	key, err := crypto.GenerateKey()
-	require.NoError(t, err)
-	return NewNode(PubkeyID(&key.PublicKey), ip, 30303, 30303, nil, NodeTypeBN)
-}
-
-//// Mock UDP transports
-
-// noopTransport is a minimal no-op transport for Table2 unit tests.
-type noopTransport struct{}
-
-func (m *noopTransport) ping(toid NodeID, toaddr *net.UDPAddr) error { return nil }
-func (m *noopTransport) waitping(from NodeID, fromIP net.IP) error   { return nil }
-func (m *noopTransport) findnode(toid NodeID, toaddr *net.UDPAddr, target NodeID, nType NodeType, max int) ([]*Node, error) {
-	return nil, nil
-}
-func (m *noopTransport) close() {}
